// src/Types/FilterFormContextStateInterface.ts
export interface FilterFormContextStateInterface {
  courseEvaluation: string;
  activityEvaluation: string;
  codeEvaluation: string;
  teamEvaluation: string;
  workplacePreference: string;
  contractPreference: string;
  salaryMin: number;
  salaryMax: number;
  internshipPreference: number | null;
  experience: number;
}